+++ telegram_listener.py
@@ -6,6 +6,7 @@ import json
 from datetime import datetime, timezone
 from config import Config
 from save_file import save_to_google_sheets
+from toto_generator import TotoGenerator
 
 class TelegramListener:
     def __init__(self):
@@ -13,6 +14,7 @@ class TelegramListener:
         self.chat_id = Config.CHAT_ID
         self.api_url = Config.get_telegram_api_url()
         self.last_update_file = Config.LAST_UPDATE_FILE
+        self.generator = TotoGenerator()
 
     def get_last_update_id(self):
         """Get the last processed update ID"""
@@ -104,97 +106,41 @@ class TelegramListener:
             return 1 <= num <= 10
         return False
 
-    def parse_generated_output(self, output):
-        """Parse the output from toto-generator.py to extract formatted strings from embedded JSON"""
+    def format_telegram_message(self, toto_data):
+        """Format TOTO data as Telegram message"""
+        message = f"🎲 *Your TOTO Numbers*\n"
+        message += f"📅 Date: {toto_data['date']}\n"
+        message += f"🎯 Total Sets: {toto_data['total_sets']}\n\n"
+
+        for set_data in toto_data['sets']:
+            message += f"*Set {set_data['set']}:* `{set_data['formatted']}`\n"
+
+        output_txt = 'sets' if toto_data['total_sets'] > 1 else 'set'
+        message += f"\n🍀 Good luck with all {toto_data['total_sets']} {output_txt}!"
+        return message
+
+    def run_toto_generator(self, sets_count, user_id=None, message_id=None):
+        """Generate TOTO numbers and handle results"""
         try:
-            print(f"DEBUG: Full output: {output}")
+            print(f"Generating {sets_count} sets of TOTO numbers...")
 
-            # Find the start of JSON after "Generated data:"
-            start_marker = "Generated data:"
-            start_pos = output.find(start_marker)
+            # Generate numbers directly (no subprocess!)
+            result = self.generator.generate_multiple_sets(int(sets_count))
 
-            if start_pos == -1:
-                print("DEBUG: No 'Generated data:' marker found")
-                return []
+            # Format and send to Telegram
+            message = self.format_telegram_message(result)
+            self.send_response(message, reply_to_message_id=message_id)
 
-            # Find the JSON starting from after the marker
-            json_start = output.find('{', start_pos)
-            if json_start == -1:
-                print("DEBUG: No opening brace found")
-                return []
+            # Save to Google Sheets if we have user info
+            if user_id and message_id:
+                formatted_sets = [set_data['formatted'] for set_data in result['sets']]
+                save_to_google_sheets(formatted_sets, user_id, message_id)
 
-            # Find the matching closing brace
-            brace_count = 0
-            json_end = json_start
+            print("TOTO generation completed successfully")
+            return True
 
-            for i in range(json_start, len(output)):
-                if output[i] == '{':
-                    brace_count += 1
-                elif output[i] == '}':
-                    brace_count -= 1
-                    if brace_count == 0:
-                        json_end = i + 1
-                        break
-
-            # Extract the complete JSON
-            json_text = output[json_start:json_end]
-            #print(f"DEBUG: Extracted JSON: {json_text}")
-
-            data = json.loads(json_text)
-
-            formatted_sets = []
-            if 'sets' in data:
-                for set_data in data['sets']:
-                    if 'formatted' in set_data:
-                        formatted_sets.append(set_data['formatted'])
-
-           # print(f"DEBUG: Extracted formatted sets: {formatted_sets}")
-            return formatted_sets
-
-        except json.JSONDecodeError as e:
-            print(f"Error parsing JSON: {e}")
-            print(f"Attempted to parse: {json_text if 'json_text' in locals() else 'No JSON extracted'}")
-            return []
         except Exception as e:
-            print(f"Error parsing generator output: {e}")
-            return []
-
-    def run_toto_generator(self, sets_count, user_id = None, message_id = None):
-        """Run the toto-generator.py with the specified number of sets"""
-        try:
-            print(f"Running TOTO generator with {sets_count} sets...")
-
-            # the Toto Generator is being run here
-            result = subprocess.run(
-                ['python', 'toto-generator.py', sets_count],
-                capture_output=True,
-                text=True,
-                timeout=60
-            )
-
-            if result.returncode == 0:
-                print("TOTO generator completed successfully")
-                # If we have user info, also save to our tracking CSV
-                if user_id and message_id:
-                    # Parse the output from toto-generator.py to extract numbers
-                    generated_sets = self.parse_generated_output(result.stdout)
-                    print("The generated sets are: ", generated_sets)
-                    if generated_sets:
-                        # Save to local CSV (optional backup)
-                    #    self.save_generated_numbers(generated_sets, user_id, message_id)
-
-                        # Save to Google Sheets
-                        save_to_google_sheets(generated_sets, user_id, message_id)
-                return True
-            else:
-                print(f"TOTO generator failed: {result.stderr}")
-                return False
-
-        except subprocess.TimeoutExpired:
-            print("TOTO generator timed out")
-            return False
-        except Exception as e:
-            print(f"Error running TOTO generator: {e}")
+            print(f"Error generating TOTO numbers: {e}")
             return False
 
     def process_telegram_messages(self):
@@ -222,8 +168,8 @@ class TelegramListener:
             message_id = message.get('message_id')
             message_date = message.get('date', 0)
             user = message.get('from', {})
-            user_id = str(user.get('id', 'Unknown'))  # Get user ID instead of name
-            user_name = user.get('first_name', 'User')  # Keep for display purposes
+            user_id = str(user.get('id', 'Unknown'))
+            user_name = user.get('first_name', 'User')
 
             if not message_text:
                 continue
@@ -254,10 +200,10 @@ class TelegramListener:
                 help_text = """TOTO Generator Bot
 
 Send me a number from 1 to 10:
-• 1 → 1 set of numbers
-• 2 → 2 sets of numbers  
-• 3 → 3 sets of numbers
-• 4 → 4 sets of numbers
+- 1 → 1 set of numbers
+- 2 → 2 sets of numbers  
+- 3 → 3 sets of numbers
+- 4 → 4 sets of numbers
 ...and so on up to 10"""
 
                 self.send_response(help_text, reply_to_message_id=message_id)
@@ -272,12 +218,10 @@ def main():
         sys.exit(1)
 
     listener = TelegramListener()
-
     had_messages = listener.process_telegram_messages()
 
     if not had_messages:
         print("No TOTO requests found - doing nothing (check only mode)")
-        # NO fallback generation here
 
 
 if __name__ == "__main__":
